pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bondsai"
version = "0.1.0"
description = "A modular AI chatbot with GPT-4o-mini integration"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "BondsAI Team", email = "team@bondsai.com"}
]
keywords = ["ai", "chatbot", "gpt", "openai", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[project.scripts]
bondsai = "bondsai.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bondsai"]

[tool.hatch.build]
packages = ["src/bondsai"]

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
